fix(body-link): address Phase 1 code review issues (#1-4)
- Issue 1: Fix serial race condition - explicitly disable Reeltwo
  with setStream(nullptr, nullptr) when body link enabled
- Issue 2: Add buffer overflow logging using DEBUG_PRINTLN
- Issue 3: Fix millis() overflow bug in health API with (int32_t) cast
- Issue 4: Increase buffer to 65 bytes, use sizeof(sBuf)-1 pattern

Build: RAM 18.3%, Flash 80.6%

+++ AsyncWebInterface.h
@@ -649,7 +649,7 @@ static String buildHealthJson()
     json += ",\"body_link\":{";
     json += "\"enabled\":" + String(bodyLinkPrefEnabled ? "true" : "false");
     json += ",\"connected\":" + String(bodyLinkConnected() ? "true" : "false");
-    json += ",\"last_rx_ms\":" + String(sBodyLastSeenMs > 0 ? (millis() - sBodyLastSeenMs) : 0);
+    json += ",\"last_rx_ms\":" + String(sBodyLastSeenMs > 0 ? (int32_t)(millis() - sBodyLastSeenMs) : 0);
     json += ",\"hb_rx\":" + String(sBodyHeartbeatRx);
     json += "}";
